Instance weights were being cast to ints for some reason. Removed the casts
git-svn-id: https://svn.cms.waikato.ac.nz/svn/weka/branches/book2ndEd-branch@4099 e0a1b77d-ad91-4216-81b1-defd5f83fa92

+++ weka/classifiers/bayes/AODE.java
@@ -51,7 +51,7 @@ import weka.classifiers.UpdateableClassifier;
  * Specify the frequency limit for parent attributes.<p>
  *
  * @author Janice Boughton (jrbought@csse.monash.edu.au) & Zhihai Wang (zhw@csse.monash.edu.au)
- * @version $Revision: 1.8.2.3 $
+ * @version $Revision: 1.8.2.4 $
  *  this version resolves errors in the handling of missing attribute values.
  */
 public class AODE extends Classifier
@@ -230,7 +230,7 @@ public class AODE extends Classifier
        return;   // ignore instances with missing class
 
     int classVal = (int)instance.classValue();
-    int weight = (int)instance.weight();
+    double weight = instance.weight();
  
     m_ClassCounts[classVal] += weight;
     m_SumInstances += weight;
